Fixed casts related to item amount [changed WORD -> DWORD] A single item stack can have up to 65535 amount (WORD) but functions like NEWGOLD/CONSUME/... can create/consume many stacks, where the total value is > 65535, so these functions must be changed to DWORD

+++ src/graysvr/CCharNPCPet.cpp
@@ -703,7 +703,7 @@ bool CChar::NPC_OnHirePay( CChar * pCharSrc, CItemMemory * pMemory, CItem * pGol
 		}
 
 		// Put all my loot cash away.
-		ContentConsume( RESOURCE_ID(RES_TYPEDEF,IT_GOLD), USHRT_MAX, false, 0 );
+		ContentConsume( RESOURCE_ID(RES_TYPEDEF,IT_GOLD), ULONG_MAX, false, 0 );
 		// Mark all my stuff ATTR_OWNED - i won't give it away.
 		ContentAttrMod( ATTR_OWNED, true );
